Cleanup the saga and the outbox operations to enable metadata sharing
