+++ src/NServiceBus.Persistence.CosmosDB/SynchronizedStorage/Operation.cs
@@ -5,6 +5,7 @@
     using Extensibility;
     using Microsoft.Azure.Cosmos;
     using Newtonsoft.Json;
+    using Newtonsoft.Json.Linq;
 
     abstract class Operation : IDisposable
     {
@@ -46,6 +47,37 @@
 
         public abstract void Apply(TransactionalBatch transactionalBatch, PartitionKeyPath partitionKeyPath);
         public virtual void Dispose() { }
+
+        protected void EnrichWithPartitionKeyIfNecessary(JObject toBeEnriched, PartitionKeyPath partitionKeyPath)
+        {
+            var partitionKeyAsJArray = JArray.Parse(PartitionKey.ToString())[0];
+            // we should probably optimize this a bit and the result might be cacheable but let's worry later
+            var pathToMatch = partitionKeyPath.ToString().Replace("/", ".");
+            var segments = pathToMatch.Split(new[] { "." }, StringSplitOptions.RemoveEmptyEntries);
+
+            var start = new JObject();
+            var current = start;
+            for (var i = 0; i < segments.Length; i++)
+            {
+                var segmentName = segments[i];
+
+                if (i == segments.Length - 1)
+                {
+                    current[segmentName] = partitionKeyAsJArray;
+                    continue;
+                }
+
+                current[segmentName] = new JObject();
+                current = (JObject)current[segmentName];
+            }
+
+            // promote it if not there, what if the user has it and the key doesn't match?
+            var matchToken = toBeEnriched.SelectToken(pathToMatch);
+            if (matchToken == null)
+            {
+                toBeEnriched.Merge(start);
+            }
+        }
     }
 
     class ThrowOnConflictOperation : Operation
